Add secret_read and rename create_secret
