+++ packs/kubernetes/actions/secret_create.yaml
@@ -1,6 +1,6 @@
 ---
-  name: "create_secret"
-  entry_point: "create_secret.py"
+  name: "secret_create"
+  entry_point: "secret_create.py"
   pack: "kubernetes"
   description: "create a kubernetes secret in a namespace"
   enabled: true
