+++ packs/kubernetes/actions/secret_read.yaml
@@ -1,0 +1,18 @@
+---
+  name: "secret_read"
+  entry_point: "secret_read.py"
+  pack: "kubernetes"
+  description: "read a kubernetes secret in a namespace"
+  enabled: true
+  runner_type: run-python
+  parameters:
+    name:
+      type: "string"
+      required: true
+      description: "secret name"
+      position: 0
+    ns:
+      type: "string"
+      description: "target namespace"
+      default: default
+      position: 0
